Move `git` to module `gardenlinux`
