+++ src/gardenlinux/git/__init__.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 from .git import Git
 
 __all__ = ["Git"]
